+++ src/db/models/station_item.rs
@@ -1,3 +1,6 @@
+use chrono::DateTime;
+use chrono::Utc;
+
 #[derive(Clone,Debug)]
 pub struct StationItem {
     pub id: i32,
@@ -16,16 +19,22 @@ pub struct StationItem {
     pub languagecodes: String,
     pub votes: i32,
     pub lastchangetime: String,
+    pub lastchangetime_iso8601: Option<DateTime<Utc>>,
     pub codec: String,
     pub bitrate: u32,
     pub hls: bool,
     pub lastcheckok: bool,
     pub lastchecktime: String,
+    pub lastchecktime_iso8601: Option<DateTime<Utc>>,
     pub lastcheckoktime: String,
+    pub lastcheckoktime_iso8601: Option<DateTime<Utc>>,
     pub lastlocalchecktime: String,
+    pub lastlocalchecktime_iso8601: Option<DateTime<Utc>>,
     pub clicktimestamp: String,
+    pub clicktimestamp_iso8601: Option<DateTime<Utc>>,
     pub clickcount: u32,
     pub clicktrend: i32,
+    pub ssl_error: bool,
     pub geo_lat: Option<f64>,
     pub geo_long: Option<f64>,
 }